Pre-Loading data
Moved port usage from ISR to main loop.  The main loop now sets the port data in advance as the Z80 is now being halted after each IN.  This work really well and not seeing any data issue.  Have a out by 1 bug currently as the loop does not know about the last item, will fix next.  Getting this in now as it's tested works .