RETN gone!
Messed up the previous test—forgot to jumper the upper ROM, leaving it at 0V. I replaced the final RETN with a self-modifying JP, which NOW works. This allows me to use HALT (via maskable), as the return stack only affects the PC value already stored by the JP back to the program. The interrupt is just RETI, so the instruction seems to execute in time before the ROM swaps out.